src/adapters/requirements.txt
numpy==1.24.2
pandas==1.5.3
psycopg2==2.9.10
psycopg2_binary==2.9.10
python_bcrypt==0.3.2
